# docs/persona-ai-mvp.openapi.yaml
openapi: 3.0.3
info:
  title: TalkTo Persona AI MVP API
  version: 0.1.0
  description: |
    Backend MVP API design for Persona AI.

    v1.0 source priority: TalkTo_개발자용_기능명세서_v1.0.md overrides older Notion rows,
    pasted planning notes, and the legacy MVP baseline where they conflict.

    Current implemented surface: legacy Persona Chat MVP plus Archive-first
    subjects, purpose consents, question cards/interactions, recording upload,
    recording completion, recording list/detail, original recording playback,
    payment products/orders/webhooks/entitlements, and queued paid Archive
    analysis jobs, analysis job status APIs, audited admin retry hooks, and
    admin worker transition endpoints for paid analysis jobs, and paid Memories
    status/search plus segment playback/feedback over indexed analysis segments,
    in-app analysis status notifications, and R&D redacted export opt-in/preview
    guardrails, plus Voice Persona application core routes.

    Guardrails: free Archive upload/completion must not create Preview, STT,
    LLM summary, embedding, memory segment, or analysis job work. Preview/sample
    analysis is deferred/disabled by v1.0. POST
    /api/v1/recordings/{recordingId}/preview-analysis exists only as a disabled
    compatibility endpoint returning 409 feature_deferred and must not enqueue
    Preview, free AI, or analysis job work.

    Not yet implemented: automated provider execution for analysis workers,
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/memories/status:
    get:
      tags: [Memories]
      summary: Get paid Memories search readiness
      description: Returns active Memories entitlement state and whether indexed analysis memory segments are available for paid search. This is separate from legacy memory CRUD.
      parameters:
        - in: query
          name: subjectId
          schema:
            type: string
            format: uuid
          description: Restrict index readiness counts to one subject.
      responses:
        '200':
          description: Paid Memories readiness status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoriesStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/memories/search:
    post:
      tags: [Memories]
      summary: Search paid Memories over analyzed Archive segments
      description: Requires an active Memories entitlement and required Memories purpose consent. Searches owned memory segments created by completed paid analysis and optional stored embeddings; it does not search legacy memory CRUD rows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoriesSearchRequest'
      responses:
        '200':
          description: Paid Memories search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoriesSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/data-rights/research-export:
    get:
      tags: [Data Rights]
      summary: Get research export opt-in preference
      responses:
        '200':
          description: Current preference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchExportPreferenceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/data-rights/research-export/opt-in:
    post:
      tags: [Data Rights]
      summary: Opt in to redacted research export eligibility
      responses:
        '201':
          description: Opt-in preference saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchExportPreferenceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/data-rights/research-export/withdraw:
    post:
      tags: [Data Rights]
      summary: Withdraw research export opt-in
      responses:
        '201':
          description: Opt-in withdrawn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchExportPreferenceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/data-deletion-requests:
    post:
      tags: [Data Rights]
      summary: Create a strongly confirmed data deletion request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataDeletionRequest'
      responses:
        '201':
          description: Data deletion request created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/data-rights/research-export-preview:
    get:
      tags: [Admin, Data Rights]
      summary: Preview redacted research export eligible rows
      description: Dry-run preview for admin/ops users. Returns only opted-in memory segment identifiers, timestamps, eligibility state, and redacted text when a successful redaction exists. Raw audio keys, presigned URLs, raw transcripts, and raw memory text are not returned.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: subjectId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Redacted export preview rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchExportPreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/data-deletion-requests/{requestId}:
    patch:
      tags: [Admin, Data Rights]
      summary: Update data deletion request status
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDataDeletionRequest'
      responses:
        '200':
          description: Deletion request updated
  /api/v1/admin/provider-deletion-records/{recordId}:
    patch:
      tags: [Admin, Data Rights]
      summary: Update provider deletion tracking status
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProviderDeletionRecord'
      responses:
        '200':
          description: Provider deletion record updated
...
  /api/v1/memories:
    get:
      tags: [Memories]
      summary: List memories
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: person
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
      responses:
        '200':
          description: Memory list
    post:
      tags: [Memories]
      summary: Create memory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryWriteRequest'
      responses:
        '201':
          description: Memory created
  /api/v1/memories/{memoryId}:
    get:
      tags: [Memories]
      summary: Get memory detail
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: Memory detail
    patch:
      tags: [Memories]
      summary: Update memory
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryWriteRequest'
      responses:
        '200':
          description: Memory updated
  /api/v1/memories/{memoryId}/deactivate:
    post:
      tags: [Memories]
      summary: Deactivate memory
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: Memory deactivated
  /api/v1/memories/{memoryId}/reembed:
    post:
      tags: [Memories]
      summary: Request memory re-embedding
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '202':
          description: Re-embedding requested
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or token is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  parameters:
...
    MemoryId:
      in: path
      name: memoryId
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: CONFLICT
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
...
    MemoriesStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MemoriesStatus'
    MemoriesStatus:
      type: object
      properties:
        entitled:
          type: boolean
        searchable:
          type: boolean
        indexedSegmentCount:
          type: integer
          minimum: 0
        completedJobCount:
          type: integer
          minimum: 0
        latestCompletedAt:
          type: string
          format: date-time
          nullable: true
    MemoriesSearchRequest:
      type: object
      required: [q]
      properties:
        q:
          type: string
          minLength: 1
          maxLength: 500
        subjectId:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        limit:
          type: integer
          minimum: 1
          maximum: 20
          default: 5
    MemoriesSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MemoriesSearchResult'
    MemoriesSearchResult:
      type: object
      properties:
        status:
          type: string
          enum: [answered, no_results]
        answer:
          type: string
        resultCount:
          type: integer
          minimum: 0
        segments:
          type: array
          items:
            $ref: '#/components/schemas/MemorySearchSegment'
        lowConfidence:
          type: boolean
    MemorySearchSegment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        startMs:
          type: integer
          minimum: 0
        endMs:
          type: integer
          minimum: 0
        speakerLabel:
          type: string
          nullable: true
        text:
          type: string
        score:
          type: number
        lowConfidence:
          type: boolean
...
    ResearchExportPreferenceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ResearchExportPreference'
    ResearchExportPreference:
      type: object
      properties:
        optedIn:
          type: boolean
        optedInAt:
          type: string
          format: date-time
          nullable: true
        withdrawnAt:
          type: string
          format: date-time
          nullable: true
    ResearchExportPreviewResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                $ref: '#/components/schemas/ResearchExportPreviewRow'
    ResearchExportPreviewRow:
      type: object
      properties:
        memorySegmentId:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        startMs:
          type: integer
          minimum: 0
        endMs:
          type: integer
          minimum: 0
        eligible:
          type: boolean
        redactedText:
          type: string
          nullable: true
        exclusionReason:
          type: string
          nullable: true
    CreateDataDeletionRequest:
      type: object
      required: [scope, confirmation]
      properties:
        scope:
          type: string
          enum: [account, subject, recording]
        confirmation:
          type: string
          example: DELETE MY DATA
        subjectId:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 500
    UpdateDataDeletionRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [requested, in_review, processing, completed, rejected]
    UpdateProviderDeletionRecord:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending_manual, requested, completed, failed]
        notes:
          type: string
          maxLength: 500
...
    MemoryWriteRequest:
      type: object
      required: [title, memory_type, body_markdown]
      properties:
        title:
          type: string
        memory_type:
          type: string
        related_people:
          type: array
          items:
            type: string
        related_period:
          type: string
          nullable: true
        body_markdown:
          type: string
        tags:
          type: array
          items:
            type: string
        confidence_score:
          type: number
          format: float
...